+++ Content.IntegrationTests/PoolManagerTestEventHandler.cs
@@ -3,28 +3,12 @@ namespace Content.IntegrationTests;
 [SetUpFixture]
 public sealed class PoolManagerTestEventHandler
 {
-    // This value is completely arbitrary.
-    private static TimeSpan MaximumTotalTestingTimeLimit => TimeSpan.FromMinutes(30);
-    private static TimeSpan HardStopTimeLimit => MaximumTotalTestingTimeLimit.Add(TimeSpan.FromMinutes(1));
+    // Goobstation edit - migrated the code to PoolManagerHelpers
 
     [OneTimeSetUp]
     public void Setup()
     {
-        PoolManager.Startup();
-        // If the tests seem to be stuck, we try to end it semi-nicely
-        _ = Task.Delay(MaximumTotalTestingTimeLimit).ContinueWith(_ =>
-        {
-            // This can and probably will cause server/client pairs to shut down MID test, and will lead to really confusing test failures.
-            TestContext.Error.WriteLine($"\n\n{nameof(PoolManagerTestEventHandler)}: ERROR: Tests are taking too long. Shutting down all tests. This may lead to weird failures/exceptions.\n\n");
-            PoolManager.Shutdown();
-        });
-
-        // If ending it nicely doesn't work within a minute, we do something a bit meaner.
-        _ = Task.Delay(HardStopTimeLimit).ContinueWith(_ =>
-        {
-            var deathReport = PoolManager.DeathReport();
-            Environment.FailFast($"Tests are TAKING TOO LONG!\n Death Report:\n{deathReport}");
-        });
+        PoolManagerHelpers.Setup();
     }
 
     [OneTimeTearDown]
